pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metasift"
version = "0.1.0"
description = "AI-powered metadata analyst and steward for OpenMetadata"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "Hero" }]
dependencies = [
    # OpenMetadata integration
    "openmetadata-ingestion==1.12.4",
    "data-ai-sdk[langchain]",

    # Agent framework
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langchain-google-genai>=2.0",

    # Data + analytics
    "duckdb>=1.1",
    "pandas>=2.2",
    "plotly>=5.24",
    "thefuzz[speedup]>=0.22",

    # UI
    "streamlit>=1.36",

    # Utilities
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pydantic>=2.7,<2.12",
    "tenacity>=9.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ipython>=8.0",
]

[tool.setuptools]
packages = ["app", "app.clients", "app.engines", "app.ui", "app.utils"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
